# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/usuarios-internos": {
      "post": {
        "operationId": "UsuariosInternosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Usuarios Internos"
        ]
      },
      "get": {
        "operationId": "UsuariosInternosController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuarios Internos"
        ]
      }
    },
    "/usuarios-internos/{id}": {
      "get": {
        "operationId": "UsuariosInternosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuarios Internos"
        ]
      },
      "patch": {
        "operationId": "UsuariosInternosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuarios Internos"
        ]
      }
    },
    "/usuarios-internos/mi-perfil": {
      "get": {
        "operationId": "UsuariosInternosController_getPerfil",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuarios Internos"
        ]
      },
      "patch": {
        "operationId": "UsuariosInternosController_updatePerfil",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePerfilDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuarios Internos"
        ]
      }
    },
    "/roles": {
      "post": {
        "operationId": "RolesController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRolDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "get": {
        "operationId": "RolesController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/{idRol}": {
      "get": {
        "operationId": "RolesController_obtener",
        "parameters": [
          {
            "name": "idRol",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_actualizar",
        "parameters": [
          {
            "name": "idRol",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_eliminar",
        "parameters": [
          {
            "name": "idRol",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/{idRol}/permisos": {
      "post": {
        "operationId": "RolesController_asignarPermisos",
        "parameters": [
          {
            "name": "idRol",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsignarPermisosDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/permisos": {
      "get": {
        "operationId": "PermisosController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Permisos"
        ]
      },
      "post": {
        "operationId": "PermisosController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermisoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Permisos"
        ]
      }
    }
  },
  "info": {
    "title": "Sistema de Gestión",
    "description": "Módulo de Gestión de Usuarios, Roles y Permisos",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "correo": {
            "type": "string"
          },
          "contraseña": {
            "type": "string"
          }
        },
        "required": [
          "correo",
          "contraseña"
        ]
      },
      "CreateUsuarioDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "contraseña": {
            "type": "string"
          },
          "idRol": {
            "type": "string"
          }
        },
        "required": [
          "nombre",
          "correo",
          "contraseña",
          "idRol"
        ]
      },
      "UpdateUsuarioDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "idRol": {
            "type": "string"
          }
        }
      },
      "UpdatePerfilDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "nuevaContraseña": {
            "type": "string"
          }
        }
      },
      "CreateRolDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          }
        },
        "required": [
          "nombre"
        ]
      },
      "UpdateRolDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          }
        }
      },
      "AsignarPermisosDto": {
        "type": "object",
        "properties": {
          "permisosIds": {
            "example": [
              "uuid_permiso_1",
              "uuid_permiso_2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permisosIds"
        ]
      },
      "CreatePermisoDto": {
        "type": "object",
        "properties": {
          "modulo": {
            "type": "string"
          },
          "accion": {
            "type": "string"
          }
        },
        "required": [
          "modulo",
          "accion"
        ]
      }
    }
  }
}